Docs: Session C context log; G42 ASPD decimal display
context_log.md: Session C entry (~100% ctx, hit limit before commit phase).
G42: StatusData.aspd int→float; status_calculator uses /10 not //10;
derived_section formats as {:.1f} (e.g. 185.3 instead of 185).

+++ core/models/status.py
@@ -17,7 +17,7 @@ class StatusData:
     hit: int = 0
     flee: int = 0
     flee2: int = 0
-    aspd: int = 0
+    aspd: float = 0.0
     max_hp: int = 0
     max_sp: int = 0
